+++ backend/app/middleware/validateModuleAccess.middleware.js
@@ -1,63 +1,51 @@
-const { ROLES } = require("../enums/roles.enum");
-const { FarmUser, Module, Farm } = require('../../models');
+const { Farm, User } = require('../../models');
 const ApiResponse = require("../utils/apiResponse");
+const { Op } = require('sequelize');
 
 class ValidateModuleAccessMiddleware {
 
-    async validate(req, res, next) {
+    async validateOwnerFarmAccess(req, res, next) {
         try {
-            const moduleId = parseInt(req.params.id);
             const authenticatedUser = req.user;
 
-            const module = await Module.findByPk(moduleId, {
+            const userFarms = await Farm.findAll({
                 include: [
                     {
-                        model: Farm,
-                        as: 'farm',
-                        attributes: ['id', 'name']
+                        model: User,
+                        as: 'users',
+                        required: true,
+                        where: {
+                            [Op.or]: [
+                                { id: authenticatedUser.id }
+                            ]
+                        },
+                        through: {
+                            attributes: []
+                        }
                     }
                 ]
             });
 
-            if (!module) {
-                return res.status(404).json(
-                    ApiResponse.createApiResponse('Module not found',
-                        [],
-                        [{
-                            msg: 'The requested module was not found'
-                        }]
-                    )
-                );
-            }
-
-            const farmUserAssociation = await FarmUser.findOne({
-                where: {
-                    id_user: authenticatedUser.id,
-                    id_farm: module.id_farm
-                }
-            });
-
-            if (!farmUserAssociation || authenticatedUser.id_rol !== ROLES.OWNER) {
+            if (!userFarms || userFarms.length === 0) {
                 return res.status(403).json(
                     ApiResponse.createApiResponse('Authorization failed',
                         [],
                         [{
-                            msg: 'You do not have permission to access modules in this farm'
+                            msg: 'No farms with associated monitor users were found.'
                         }]
                     )
                 );
             }
 
-            req.module = module;
-
+            req.userFarms = userFarms;
             next();
         } catch (error) {
-            console.error('Module access validation error:', error);
+            console.error('Owner farm access validation error:', error);
             return res.status(500).json(
                 ApiResponse.createApiResponse('Server error',
                     [],
                     [{
-                        msg: 'Error validating permissions to access module'
+                        msg: 'Error validating access permissions'
                     }]
                 )
             );
