+++ backend/app/routes/owner/user.owner.route.js
@@ -7,7 +7,7 @@ const UserOwnerController = require('../../controllers/owner/user.owner.controll
 const ValidateTokenMiddleware = require('../../middleware/validateToken.middleware');
 const BlackListService = require('../../services/shared/blacklist.service');
 const UserService = require("../../services/shared/user.service");
-const ValidateModuleAccessMiddleware = require("../../middleware/validateModuleAccessMiddleware");
+const ValidateModuleAccessMiddleware = require("../../middleware/validateModuleAccess.middleware");
 
 const { ROLES: Role } = require("../../enums/roles.enum");
 const ValidateRoleMiddleware = require("../../middleware/validateRole.middleware");
